requirements.txt
python-telegram-bot==20.1
PyMuPDF==1.23.8
Pillow==9.5.0
numpy==1.25.0
pyyaml==6.0